jsファイルのリファクタリング
